fix: limit concurrent HTTP requests (#16)
* fix: limit concurrent HTTP requests

All HTTP requests made by this module are sent to the same delegate
host. Browsers throttle the number of concurrent requests per hostname,
right now it is 6 per host, which suffocates the use of delegate and
blocking it from being used for preload or delegated peer routing.

This change introduces task queues that limit the number of concurrent
requests, making it safe to run in browser context.

Optimizations:
- preload calls via `refs` are known to take time,
  so we limit them to max two at a time
- swarm.connect was the main offender (causing multiple requests to delegate),
  we now check if connection is already present and cache result for 1 minute
  removing most of redundant http requests
- hostname of default delegate is changed

Context: https://github.com/libp2p/js-libp2p-delegated-content-routing/issues/12
Closes #12

* refactor: remove the need for swarm connect

This removes all code that caused swarm connect calls from delegate to
self. It was not needed: delegate is one of bootstrap nodes,
so we are always connected to it.

https://github.com/libp2p/js-libp2p-delegated-content-routing/issues/12#issuecomment-514591525

* chore: remove unused dependencies

We no longer need those (removed calls to swarm.connect)

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>