feat: allow providing non-dag-pb nodes
The http endpoint we send the CID to be provided only understands
dag-pb nodes.

Change the endpoints from a non-recursive `ipfs refs` call to a
`ipfs block stat` instead as it does not try to interperet the data.

N.b. only the passed CID will be provided, not any children but since
the original call was a non-recursive `ipfs refs` the behaviour should
be the same.

Also expands the explanation of how the provide happens and adds the
number of found providers to the logging output to make debugging things
a little easier.